Moved InputStreamChunkEncode out of ChunkedBody.
